Update dependencies to include newer Cabal
The upper bound on Cabal has been increased. The lower bound on Cabal
has also been increased, to avoid a mismatch between the newest version
and the one that comes bundled with GHC: the test runner tended to
switch back to the old Cabal, but generates a stub which is only
compatible with the new one. For now, use of this package in a test
suite will require an appropriately high lower bound on Cabal.

The testing GHC options now include -Werror to act as a better testing
example, and the version number is bumped for release.