Weaken Cabal bounds, remove example test-suite
The lower bound on Cabal was unnecessary and has been relaxed. The
example test-suite, which includes failures, has been removed so that
the project does not appear to be failing its tests. The Example module
is still available, now with some extra documentation, and the readme
has been updated to include the relevant information that should be in a
Cabal file.  The source has also been tidied by stylish-haskell.