Update upper bound on base to include GHC 7.10
This change extends the upper bound on base to include anything under
version 5. Because of the changes to the prelude, the CPP is now used to
conditionally import definitions to avoid warnings. Activating the CPP
appears to affect multiline strings, so the one multiline string in the
project is now two concatenated strings.